+++ config/Migrations/20201014000224_MigrationAddresses.php
@@ -75,7 +75,7 @@ class MigrationAddresses extends AbstractMigration {
 				'precision' => 10,
 				'scale' => 6,
 			])
-			->addColumn('lng', 'float', [
+			->addColumn('lng', 'decimal', [
 				'default' => null,
 				'null' => true,
 				'precision' => 10,
