+++ config/Migrations/20201014000225_MigrationPostalCodes.php
@@ -45,7 +45,7 @@ class MigrationPostalCodes extends AbstractMigration {
 				'precision' => 10,
 				'scale' => 6,
 			])
-			->addColumn('lng', 'float', [
+			->addColumn('lng', 'decimal', [
 				'default' => null,
 				'null' => true,
 				'precision' => 10,
